+++ app/Http/Resources/Api/V1/CustomerResource.php
@@ -34,6 +34,7 @@ class CustomerResource extends JsonResource
 
         return [
             'id' => $this->resource->id,
+            'legal_entity_id' => $this->resource->legal_entity_id,
             'customer_number' => $this->resource->customer_number,
             'name' => $this->resource->name,
             'billing_address' => $this->resource->billing_address,